requirements.txt
Flask==3.0.2
Flask-Session==0.5.0
msal==1.26.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
PyMySQL==1.1.0
cryptography==42.0.7
qrcode==7.4.2
Pillow==10.2.0
requests==2.31.0